Failed_Game.js Was Reborn
This is a failed game I made in CS class. It was too complicated to be considered correct so I had to dumb it down in CS. This is the original version I created. It was too complicated for CS class I guess. I shall upload this to Github because I want to remember this game for the rest of my life and I need a place to save it.